+++ sbin/lmserv.py
@@ -84,7 +84,7 @@ eva.mailer.update_config(cfg)
 
 eva.core.load_cvars()
 
-eva.apikey.allows = ['cmd', 'dm_rules_props', 'dm_rules_list']
+eva.apikey.allows = ['cmd', 'lock', 'dm_rules_props', 'dm_rules_list']
 eva.apikey.load()
 
 eva.notify.init()